requirements.txt
aniso8601==8.0.0
bcrypt==3.2.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
DateTime==4.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.0.0
pytz==2020.1
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.26.2
uuid==1.30
Werkzeug==1.0.1
zope.interface==5.2.0
